Do not use `rb_raise` to gen exception from Go
This interacts weirdly with the GoGC and the Go control mechanism, as
this pushes the return value of a _Go_ function to a _C_ object. This
trips up the Go Gargabe collector.

I have **not** removed all usages of `rb_raise` for now, just in this
`Fetch` method, renamed it to `_fetch` and wrapped it in a `fetch`
method and doing the exception raising from the Ruby code.